delete project, edit fix, refactor check user id

+++ src/main/java/com/epitools/homer/homer/rest/controller/ProjectRestController.java
@@ -4,8 +4,8 @@ import com.epitools.homer.homer.model.Project;
 import com.epitools.homer.homer.model.User;
 import com.epitools.homer.homer.repository.ProjectRepository;
 import com.epitools.homer.homer.repository.UserRepository;
+import com.epitools.homer.homer.util.Utils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -40,18 +40,11 @@ public class ProjectRestController {
     @RequestMapping(value="/projects/{id}", method=RequestMethod.PUT, produces={ MediaType.APPLICATION_JSON_VALUE })
     public ResponseEntity<Object> updateProject(@PathVariable(value="id") Integer projectId,
                                            @Valid @RequestBody Project projectDetails) {
-        final String user = SecurityContextHolder.getContext().getAuthentication().getPrincipal().toString();
-        final User maybeUser = userRepository.findByEmail(user);
-        if (maybeUser == null) return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .contentType(MediaType.APPLICATION_JSON)
-                .body("{ \"error\" : \"User not connected\" }");
+        final User maybeUser = Utils.getMaybeUser(userRepository);
+        if (maybeUser == null) return Utils.jsonError("User not connected");
         final Project project = projectRepository.findOne(projectId);
         if(project == null) return ResponseEntity.notFound().build();
-        if (!project.getUserId().equals(maybeUser.getId())) return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .contentType(MediaType.APPLICATION_JSON)
-                .body("{ \"error\" : \"Can not edit this project\" }");
+        if (!project.getUserId().equals(maybeUser.getId())) return Utils.jsonError("Can not edit this project");
         project.setUserId(projectDetails.getUserId() == null ? project.getUserId() : projectDetails.getUserId());
         project.setSpices(projectDetails.getSpices());
         project.setCurrentSpices(projectDetails.getCurrentSpices() == null ?
@@ -70,9 +63,12 @@ public class ProjectRestController {
     }
 
     @RequestMapping(value="/projects/{id}", method=RequestMethod.DELETE, produces={ MediaType.APPLICATION_JSON_VALUE })
-    public ResponseEntity<Project> deleteProject(@PathVariable(value="id") Integer projectId) {
-        Project project = projectRepository.findOne(projectId);
+    public ResponseEntity<Object> deleteProject(@PathVariable(value="id") Integer projectId) {
+        final User maybeUser = Utils.getMaybeUser(userRepository);
+        if (maybeUser == null) return Utils.jsonError("User not connected");
+        final Project project = projectRepository.findOne(projectId);
         if(project == null) return ResponseEntity.notFound().build();
+        if (!project.getUserId().equals(maybeUser.getId())) return Utils.jsonError("Can not delete this project");
         projectRepository.delete(project);
         return ResponseEntity.ok().build();
     }
